version1,基本完成功能

+++ model/device_status.go
@@ -18,7 +18,7 @@ type DeviceStatus struct {
 	ID				uint64      `gorm:"column:id`				//用来对设备数据进行分段的标识
 	DeviceID 		uint64 		`gorm:"column:deviceid"`		//设置DeviceID对应在数据库表里的列为deviceid,设备id，用于区分设备，为0时表示未知设备
 	DeviceName  	string		`gorm:"column:devicename"`		//设备名称
-	TTRuntime		uint64		`gorm:"column:tt_runtime"`		//检测的任务的周期运行时间，单位ms
+	TTRuntime		float64		`gorm:"column:tt_runtime"`		//检测的任务的周期运行时间，单位ms
 	CpuUseRate  	float64		`gorm:"column:cpu_userate"`		//CPU 使用率
 	CpuAvailable  	float64		`gorm:"column:cpu_available"`	//CPU 可用率
 	MemUseRate  	float64		`gorm:"column:mem_userate"`		//内存 使用率
